Add the ability to specify IPv6 blocks.
This adds the ability to specify IPv6 CIDR blocks for the security group
rules in the consul modules.

This is a fairly straight forward addition, with the only item of note
is that the IPv6 CIDR block variables are always optional, with an empty
default.

This should work even for the cases where the IPv4 CIDR blocks are
explicitly empty, to allow for IPv6 only environments.  This has not yet
been tested.

These changes have not yet been run through terraform fmt, that will
happen on the next commit.